refactor: extract config module from watchmedo.py
